+++ src/DependencyInjection/Configuration.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 
 namespace Meritoo\CommonBundle\DependencyInjection;
 
+use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
@@ -27,8 +28,41 @@ class Configuration implements ConfigurationInterface
     public function getConfigTreeBuilder(): TreeBuilder
     {
         $treeBuilder = new TreeBuilder();
-        $treeBuilder->root('meritoo_common');
+        $rootNode = $treeBuilder->root('meritoo_common');
+
+        $rootNode
+            ->children()
+            ->append($this->getApplicationNode())
+            ->end();
 
         return $treeBuilder;
+    }
+
+    /**
+     * Returns the "application" node
+     *
+     * @return NodeDefinition
+     */
+    private function getApplicationNode(): NodeDefinition
+    {
+        $treeBuilder = new TreeBuilder();
+        $rootNode = $treeBuilder->root('application');
+
+        $rootNode
+            ->addDefaultsIfNotSet()
+            ->children()
+            ->arrayNode('version')
+            ->addDefaultsIfNotSet()
+            ->children()
+            ->scalarNode('file_path')
+            ->info('Path of a file who contains version of the application')
+            ->defaultValue('%kernel.project_dir%/VERSION')
+            ->cannotBeEmpty()
+            ->end()
+            ->end()
+            ->end()
+            ->end();
+
+        return $rootNode;
     }
 }