+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 23. Tests > add kernel
 24. Composer > require `symfony/phpunit-bridge` and `symfony/yaml` packages
 25. Groundwork of Symfony's service, the `BaseService` class
+26. Service > add service that serves application, the `ApplicationService` class
 
 # 0.0.1